# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carshark",
        "version": "1.0"
    },
    "paths": {
        "/api/cars": {
            "get": {
                "tags": [
                    "Машины"
                ],
                "summary": "Получить список машин",
                "description": "Получение списка всех машин",
                "operationId": "/api/cars(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список машин"
                    }
                }
            },
            "post": {
                "tags": [
                    "Машины"
                ],
                "summary": "Добавить машину",
                "description": "Создание машины",
                "operationId": "/api/cars(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создание новой машины"
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "get": {
                "tags": [
                    "Машины"
                ],
                "summary": "Получить машину",
                "description": "Получение машины по идентификатору",
                "operationId": "/api/cars/{cars}(GET)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает машину"
                    }
                }
            },
            "put": {
                "tags": [
                    "Машины"
                ],
                "summary": "Редактировать машину",
                "description": "Редактирование машины по идентификатору",
                "operationId": "/api/cars(PUT)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирование машины"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Машины"
                ],
                "summary": "Удалить машину",
                "description": "Удаление машины по идентификатору",
                "operationId": "/api/cars/{cars}(DELETE)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удаление машины"
                    }
                }
            }
        },
        "/api/car_events": {
            "get": {
                "tags": [
                    "События машины"
                ],
                "summary": "Получить список событий машин",
                "description": "Получение списка всех событий машин",
                "operationId": "/api/car_events(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список событий машин"
                    }
                }
            },
            "post": {
                "tags": [
                    "События машины"
                ],
                "summary": "Добавить событие машины",
                "description": "Создание события машины",
                "operationId": "/api/car_events(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создание нового события машины"
                    }
                }
            }
        },
        "/api/car_events/{id}": {
            "get": {
                "tags": [
                    "События машины"
                ],
                "summary": "Получить событие машины",
                "description": "Получение события машины по идентификатору",
                "operationId": "/api/car_events{car_events}(GET)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор события машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает событие машину"
                    }
                }
            },
            "put": {
                "tags": [
                    "События машины"
                ],
                "summary": "Редактировать событие машины",
                "description": "Редактирование события машины по идентификатору",
                "operationId": "/api/car_events(PUT)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор события машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarEventStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирование события машины"
                    }
                }
            },
            "delete": {
                "tags": [
                    "События машины"
                ],
                "summary": "Удалить событие машины",
                "description": "Удаление собятия машины по идентификатору",
                "operationId": "/api/car_events/{car_events}(DELETE)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор события машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удаление собятия машины"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CarEventStoreRequest": {
                "required": [
                    "date",
                    "latitude",
                    "langitude",
                    "fuel",
                    "speed",
                    "renter_id",
                    "car_id",
                    "car_status_id"
                ],
                "properties": {
                    "date": {
                        "description": "Дата и время события",
                        "type": "dateTime",
                        "example": "2016-11-08 07:05:41"
                    },
                    "latitude": {
                        "description": "Широта",
                        "type": "number",
                        "format": "float",
                        "example": "35.55242"
                    },
                    "langitude": {
                        "description": "Долгота",
                        "type": "number",
                        "format": "float",
                        "example": "23.343332"
                    },
                    "fuel": {
                        "description": "Топливо",
                        "type": "integer",
                        "example": "23"
                    },
                    "speed": {
                        "description": "Скорость",
                        "type": "integer",
                        "example": "44"
                    },
                    "renter_id": {
                        "description": "id арендатора",
                        "type": "integer",
                        "example": "3"
                    },
                    "car_id": {
                        "description": "id машины",
                        "type": "integer",
                        "example": "3"
                    },
                    "car_status_id": {
                        "description": "id статуса машины",
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            },
...
            "CarStoreRequest": {
                "required": [
                    "number",
                    "mileage",
                    "car_statuses_id",
                    "year_of_release",
                    "car_model_id"
                ],
                "properties": {
                    "number": {
                        "description": "Номер машины",
                        "type": "string",
                        "example": "ц135вы"
                    },
                    "mileage": {
                        "description": "Пробег машины",
                        "type": "integer",
                        "example": "352"
                    },
                    "car_statuses_id": {
                        "description": "id статуса машины",
                        "type": "integer",
                        "example": "2"
                    },
                    "year_of_release": {
                        "description": "Год выпуска машины",
                        "type": "integer",
                        "example": "2003"
                    },
                    "car_model_id": {
                        "description": "id модели",
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            },
...
        }
    }
}